FAST-API/app/requirements.txt
fastapi[standard]
uvicorn[standard]
psycopg2-binary
mysql-connector-python
pymysql
python-dotenv
sqlalchemy